requirements.txt
attrs==19.3.0
Automat==0.8.0
blinker==1.4
certifi==2019.11.28
chardet==3.0.4
charset-normalizer==2.0.12
click==8.0.4
cloud-init==21.4
colorama==0.4.3
command-not-found==0.3
commonmark==0.9.1
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
dbus-python==1.2.16
decorator==5.1.1
distro==1.4.0
distro-info===0.23ubuntu1
docopt==0.6.2
entrypoints==0.3
ffmpeg==1.4
ffmpy==0.3.0
filelock==3.6.0
Flask==2.0.3
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.5.0
incremental==16.10.1
itsdangerous==2.1.1
Jinja2==3.0.3
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
MarkupSafe==2.1.0
more-itertools==4.2.0
mutagen==1.45.1
netifaces==0.10.4
oauthlib==3.1.0
pexpect==4.6.0
pipreqs==0.4.11
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycryptodomex==3.14.1
pydub==0.25.1
Pygments==2.11.2
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.0+ubuntu0.20.4.7
python-debian===0.1.36ubuntu1
python-dotenv==0.19.2
PyYAML==5.3.1
requests==2.27.1
requests-file==1.5.1
requests-unixsocket==0.2.0
rich==11.2.0
savify==2.3.4
SecretStorage==2.3.1
sentry-sdk==1.5.6
service-identity==18.1.0
simplejson==3.16.0
six==1.16.0
sos==4.2
spotify-dl==8.1.0
spotipy==2.19.0
ssh-import-id==5.10
systemd-python==234
tldextract==3.2.0
Twisted==18.9.0
ubuntu-advantage-tools==27.6
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.8
validators==0.18.2
wadllib==1.3.3
websockets==10.2
Werkzeug==2.0.3
yarg==0.1.9
youtube-dl==2021.12.17
youtube-search==2.1.0
yt-dlp==2022.2.4
zipp==1.0.0
zope.interface==4.7.1
